refactor(agent): extract chat model and skill hooks
